refactor: simplify CLI code per codex review
- consolidate git root + server mode checks in pull/push
- simplify output() to single ternary
- use .some() for branch detection
- extract MOCK_PRIORITY_CONFIG to test helpers
- merge git-sync tests into args.test.ts
