+++ packages/cli/tests/args.test.ts
@@ -58,4 +58,21 @@ describe('parseArgs', () => {
     expect(result.command).toBe('project');
     expect(result.subcommand).toBe('list');
   });
+
+  it('parses pull command', () => {
+    const result = parseArgs(['pull']);
+    expect(result.command).toBe('pull');
+  });
+
+  it('parses push command with message', () => {
+    const result = parseArgs(['push', 'update tasks']);
+    expect(result.command).toBe('push');
+    expect(result.subcommand).toBe('update tasks');
+  });
+
+  it('parses push command without message', () => {
+    const result = parseArgs(['push']);
+    expect(result.command).toBe('push');
+    expect(result.subcommand).toBeUndefined();
+  });
 });
